#164 send ON_CHANGE_SUCCESS on VRR fail
