+++ tst/wrap-info.c
@@ -19,3 +19,7 @@ void __wrap_print_mode(enum LogThreshold t, struct Mode *mode) {
 	check_expected(mode);
 }
 
+void __wrap_report_adaptive_sync_fail(struct Head *head) {
+	check_expected(head);
+}
+
